+++ backend/handlers/task/migratetasktodaily.go
@@ -12,7 +12,7 @@ import (
 )
 
 type MigrateTaskToDailyLogInput struct {
-	ID   uint64 `path:"id" example:"123" doc:"ID of the task entry"`
+	ID   uint `path:"id" example:"123" doc:"ID of the task entry"`
 	Body struct {
 		Date string `json:"date" doc:"Date when the task should be snoozed"`
 	}
@@ -46,7 +46,7 @@ func MigrateTaskToDailyLogHandler(api huma.API, es *database.EntryService) {
 	})
 }
 
-func MigrateToDaily(es *database.EntryService, entryID uint64, date date.DayDate) error {
+func MigrateToDaily(es *database.EntryService, entryID uint, date date.DayDate) error {
 	t, err := es.GetEntryByID(entryID)
 	if err != nil {
 		return err
